requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.4
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
nest-asyncio==1.5.6
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
pandas==1.1.5
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.15.0
setproctitle==1.2.3
six==1.15.0
sklearn==0.0.post1
smmap==5.0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-datasets==4.5.2
tensorflow-estimator==2.6.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.6
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
